# openapi/backend.open-api.yml
openapi: 3.1.0
info:
  title: Keycloak Backend Device Enrollment & Approval API
  version: 1.0.0
  description: |
    Server-to-server API implemented by the backend and called by Keycloak.
    Covers device enrollment checks, device-user binding, approval workflow, and device registry queries.
servers:
  - url: https://backend.example.com
    description: Backend base URL
...
paths:
  /v1/enrollments/precheck:
    post:
      tags: [enrollment]
      summary: Pre-check whether a device key can be enrolled
      description: |
        Keycloak calls this to detect duplicates and enforce policy before persisting.
        Backend should check if device_id or jkt is already bound to a user, revoked, etc.
      operationId: enrollmentPrecheck
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentPrecheckRequest'
      responses:
        '200':
          description: Policy decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentPrecheckResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/enrollments/bind:
    post:
      tags: [enrollment]
      summary: Bind an enrolled device key to a user
      description: |
        Called when Keycloak has identified/created the user and is ready to persist.
        Backend should atomically bind (device_id,jkt) to user_id and store public key material.
        If already bound to the same user, should be idempotent.
      operationId: enrollmentBind
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentBindRequest'
      responses:
        '200':
          description: Bound (idempotent)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentBindResponse'
        '409':
          description: Device already bound to a different user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/approvals:
    post:
      tags: [approvals]
      summary: Create a new-device approval request
      description: |
        Called when an existing user is adding a new device and your policy requires approval.
        Backend should notify existing devices (push/poll feed) and return a request_id.
      operationId: createApproval
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Conflict (e.g., duplicate pending request)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/approvals/{request_id}:
    get:
      tags: [approvals]
      summary: Get approval request status
      description: |
        Keycloak polls this endpoint server-to-server, then serves a KC-local polling endpoint to the browser.
      operationId: getApproval
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalStatusResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags: [approvals]
      summary: Cancel an approval request
      operationId: cancelApproval
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Cancelled
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/approvals/{request_id}/decision:
    post:
      tags: [approvals]
      summary: Decide an approval request
      description: |
        Used by an approved device (or delegated UI) to approve or deny a pending approval request.
      operationId: decideApproval
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecisionRequest'
      responses:
        '200':
          description: Updated status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/users/{user_id}/approvals:
    get:
      tags: [approvals]
      summary: List approval requests for a user
      description: |
        Returns approval requests that belong to the specified user.
      operationId: listUserApprovals
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
              enum: [PENDING, APPROVED, DENIED, EXPIRED]
      responses:
        '200':
          description: Approval list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserApprovalsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/sms/send:
    post:
      tags: [enrollment]
      summary: Send an OTP via SMS
      description: |
        Called when a new device enrollment needs an OTP delivered to the user's phone.
      operationId: sendSms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmsSendRequest'
      responses:
        '200':
          description: OTP queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmsSendResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/sms/confirm:
    post:
      tags: [enrollment]
      summary: Confirm an OTP
      description: |
        Validates the OTP/hash pair that was previously issued to the user.
      operationId: confirmSms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmsConfirmRequest'
      responses:
        '200':
          description: Confirmation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmsConfirmResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/enrollments/phone/resolve:
    post:
      tags: [enrollment]
      summary: Resolve user by phone and recommended enrollment path
      description: |
        Backend resolves whether the phone belongs to an existing user and whether the user has active devices.
        This lets Keycloak keep routing logic minimal.
      operationId: resolveUserByPhone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneResolveRequest'
      responses:
        '200':
          description: Phone resolution and routing recommendation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneResolveResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/enrollments/phone/resolve-or-create:
    post:
      tags: [enrollment]
      summary: Resolve existing user by phone or create one
      description: |
        Used after OTP verification to offload user find/create behavior to backend.
      operationId: resolveOrCreateUserByPhone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneResolveOrCreateRequest'
      responses:
        '200':
          description: User resolved or created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneResolveOrCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: |
        If provided, backend should ensure repeated calls with the same key are idempotent.
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Rate limited
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    EnrollmentPrecheckRequest:
      type: object
      required: [realm, client_id, device_id, jkt]
      properties:
        realm:
          type: string
          example: my-realm
        client_id:
          type: string
          example: mobile-app
        user_hint:
          type: string
          description: Optional user identifier hint (phone/email) if known at precheck time
          example: +491234567890
        device_id:
          type: string
          example: 3f6a0f2b-1f2e-4c37-9f7e-10f1a7c2b0d1
        jkt:
          type: string
          description: JWK thumbprint (base64url)
          example: 7bYwzOQ8f0o6u9p...
        public_jwk:
          type: object
          description: Optional JWK for convenience; may be omitted at precheck
          additionalProperties: true
        proof_context:
          type: object
          additionalProperties: true
          description: |
            Free-form context: ip, user-agent, platform, etc.
    EnrollmentPrecheckResponse:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [ALLOW, REQUIRE_APPROVAL, REJECT]
        reason:
          type: string
          example: device_already_bound
        bound_user_id:
          type: string
          nullable: true
          description: If known, the user_id the device is already bound to
        retry_after_seconds:
          type: integer
          nullable: true
    EnrollmentBindRequest:
      type: object
      required: [realm, client_id, user_id, device_id, jkt, public_jwk]
      properties:
        realm:
          type: string
        client_id:
          type: string
        user_id:
          type: string
          description: Keycloak internal user id
        user_hint:
          type: string
          description: Optional username (e.g., phone E.164)
        device_id:
          type: string
        jkt:
          type: string
        public_jwk:
          type: object
          additionalProperties: true
        attributes:
          type: object
          additionalProperties:
            type: string
          description: Extra attributes to store with the device
        created_at:
          type: string
          format: date-time
        proof:
          type: object
          description: Optional proof summary (ts, nonce, sig hash) for audit
          additionalProperties: true
    EnrollmentBindResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [BOUND, ALREADY_BOUND]
        device_record_id:
          type: string
          nullable: true
        bound_user_id:
          type: string
    ApprovalCreateRequest:
      type: object
      required: [realm, client_id, user_id, new_device]
      properties:
        realm:
          type: string
        client_id:
          type: string
        user_id:
          type: string
        new_device:
          $ref: '#/components/schemas/DeviceDescriptor'
        reason:
          type: string
          example: add_new_device
        expires_at:
          type: string
          format: date-time
        context:
          type: object
          additionalProperties: true
          description: ip, user agent, geo hints, etc.
    ApprovalCreateResponse:
      type: object
      required: [request_id, status]
      properties:
        request_id:
          type: string
          example: apr_01HZX8V9A3M1Y8S9R6W3B4KQ2P
        status:
          type: string
          enum: [PENDING]
        expires_at:
          type: string
          format: date-time
    ApprovalStatusResponse:
      type: object
      required: [request_id, status]
      properties:
        request_id:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, DENIED, EXPIRED]
        decided_at:
          type: string
          format: date-time
          nullable: true
        decided_by_device_id:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
    ApprovalDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [APPROVE, DENY]
        decided_by_device_id:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
    UserApprovalRecord:
      type: object
      required: [request_id, user_id, device_id, status, created_at]
      properties:
        request_id:
          type: string
        user_id:
          type: string
        device_id:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, DENIED, EXPIRED]
        created_at:
          type: string
          format: date-time
        decided_at:
          type: string
          format: date-time
          nullable: true
        decided_by_device_id:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
    UserApprovalsResponse:
      type: object
      required: [user_id, approvals]
      properties:
        user_id:
          type: string
        approvals:
          type: array
          items:
            $ref: '#/components/schemas/UserApprovalRecord'
    DeviceDescriptor:
      type: object
      required: [device_id, jkt]
      properties:
        device_id:
          type: string
        jkt:
          type: string
        public_jwk:
          type: object
          additionalProperties: true
          nullable: true
        platform:
          type: string
          nullable: true
          example: ios
        model:
          type: string
          nullable: true
        app_version:
          type: string
          nullable: true
...
    SmsSendRequest:
      type: object
      required: [realm, client_id, phone_number, otp]
      properties:
        realm:
          type: string
        client_id:
          type: string
        user_id:
          type: string
          nullable: true
        phone_number:
          type: string
        otp:
          type: string
        session_id:
          type: string
          nullable: true
        trace_id:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
    SmsSendResponse:
      type: object
      required: [hash]
      properties:
        hash:
          type: string
        ttl_seconds:
          type: integer
          nullable: true
        status:
          type: string
          nullable: true
    SmsConfirmRequest:
      type: object
      required: [hash, otp]
      properties:
        hash:
          type: string
        otp:
          type: string
    SmsConfirmResponse:
      type: object
      required: [confirmed]
      properties:
        confirmed:
          type: boolean
        reason:
          type: string
          nullable: true
    EnrollmentPath:
      type: string
      enum: [APPROVAL, OTP]
    PhoneResolveRequest:
      type: object
      required: [realm, client_id, phone_number]
      properties:
        realm:
          type: string
        client_id:
          type: string
        phone_number:
          type: string
        user_hint:
          type: string
          nullable: true
    PhoneResolveResponse:
      type: object
      required: [phone_number, user_exists, has_device_credentials, enrollment_path]
      properties:
        phone_number:
          type: string
        user_exists:
          type: boolean
        has_device_credentials:
          type: boolean
        enrollment_path:
          $ref: '#/components/schemas/EnrollmentPath'
        user_id:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
    PhoneResolveOrCreateRequest:
      type: object
      required: [realm, client_id, phone_number]
      properties:
        realm:
          type: string
        client_id:
          type: string
        phone_number:
          type: string
    PhoneResolveOrCreateResponse:
      type: object
      required: [phone_number, user_id, username, created]
      properties:
        phone_number:
          type: string
        user_id:
          type: string
        username:
          type: string
        created:
          type: boolean
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: bad_request
        message:
          type: string
          example: Invalid payload
        details:
          type: object
          additionalProperties: true
          nullable: true